BUGFIX: Windows path and cleanup code corrected
In windows, ':' cannot occur in a filename. So replacing the occurences of
':' with '_' for windows.

Also updated the cleanup code to remove the a.exe file generate on
windows.

+++ benchmark.py
@@ -63,7 +63,11 @@ def prepareCSV(data):
 
 def exportCSV(csv_string):
 	if (len(csv_string) > 0):
-		f = open(str(datetime.now())+".csv", "w")
+                out_file_name = str(datetime.now())+".csv";
+                if os.name == "nt":
+                        out_file_name = out_file_name.replace(":", "_", 2)
+                
+		f = open(out_file_name, "w")
 
 		# Write header on the CSV file
 		f.write("Filename,Round,Time Python,Time Cpp\n")
@@ -74,7 +78,11 @@ def exportCSV(csv_string):
 
 def cleanup():
 	os.system("rm -f temp.cpp")
-	os.system("rm -f a.out");
+
+	if os.name == "nt":
+                os.system("rm -f a.exe")
+        else:
+                os.system("rm -f a.out");
 
 def main():
 	input_files = sys.argv[1:]
